Title: 합과 평균의 차, Time: 115ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240904/합과 평균의 차/sub-of-average-and-sum.py
@@ -1,6 +1,6 @@
 a,b,c= map(int,input().split())
 hap=a+b+c
-average=hap/3
+average=int(hap/3)
 final=hap-average
 print(hap)
 print(average)
